+++ merge pok and daddy/Pastaga/src/PastaVoid/Step.java
@@ -8,15 +8,16 @@ package PastaVoid;
  * To change this template use File | Settings | File Templates.
  */
 public class Step {
-    public  int     x;
-    public  int     y;
-    public  int     holeSize;
+    public  float     x;
+    public  float     y;
+    public  float     holeSize;
     public boolean  isWall;
 
-    public Step(int x, int y, int holeSize, boolean isWall) {
+    public Step(float x, float y, float holeSize, boolean isWall) {
         this.x = x;
         this.y = y;
         this.holeSize = holeSize;
         this.isWall = isWall;
+        //System.out.printf("%f;%f => %f\n", this.x,this.y,this.holeSize);
     }
 }
